backend/requirements.txt
# FastAPI Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

# Redis & Caching
redis==5.0.1
hiredis==2.3.2

# Celery
celery==5.3.6
flower==2.0.1

# AI & LLM
anthropic==0.18.1
openai==1.12.0
langchain==0.1.6
chromadb==0.4.22
tiktoken==0.6.0

# Browser Automation
playwright==1.41.2
selenium==4.17.2
beautifulsoup4==4.12.3
lxml==5.1.0

# HTTP & Networking
httpx==0.26.0
aiohttp==3.9.3
requests==2.31.0

# Security & Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
bcrypt==4.1.2

# Utilities
python-dotenv==1.0.1
pydantic-core==2.14.6
pytz==2024.1
tzdata==2024.1

# Data Processing
pandas==2.2.0
numpy==1.26.3
orjson==3.9.13

# File Handling
openpyxl==3.1.2
python-magic==0.4.27
aiofiles==23.2.1

# Reporting
reportlab==4.0.9
jinja2==3.1.3
weasyprint==60.2
markdown==3.5.2

# Logging & Monitoring
loguru==0.7.2
prometheus-client==0.19.0
sentry-sdk==1.40.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.4
pytest-cov==4.1.0
httpx==0.26.0

# Code Quality
black==24.1.1
flake8==7.0.0
isort==5.13.2
mypy==1.8.0

# Validation & Parsing
validators==0.22.0
url-normalize==1.4.3
jsonschema==4.21.1
pyyaml==6.0.1

# Computer Vision (for screenshot analysis)
pillow==10.2.0
opencv-python==4.9.0.80

# Email
emails==0.6
jinja2==3.1.3

# Rate Limiting
slowapi==0.1.9

# WebSockets
websockets==12.0

# Cryptography
cryptography==42.0.2
email-validator
